pyproject.toml
[project]
name = "zerowaste-pattern"
version = "1.0.0"
description = "OpenEnv RL environment for garment marker making — The AI Sustainable Tailor"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "ZeroWaste Team"},
]
keywords = ["openenv", "reinforcement-learning", "sustainability", "garment-manufacturing", "packing"]

dependencies = [
    "numpy>=1.21.0",
    "shapely>=1.8.0",
    "gymnasium>=0.26.0",
    "pydantic>=2.0.0",
    "matplotlib>=3.4.0",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "openai>=1.0.0",
    "openenv-core>=0.2.0",
]

[project.optional-dependencies]
train = ["stable-baselines3>=2.0.0", "sb3-contrib>=2.0.0"]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["env*", "models*", "tasks*", "utils*", "server*"]
